% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiator.R
\name{new_radiator}
\alias{new_radiator}
\title{Create a Radiator Object}
\usage{
new_radiator(name, nominal_power, exponent = 1.3)
}
\arguments{
\item{name}{Character name of the radiator (e.g., "Living Room")}

\item{nominal_power}{Watts at Delta T = 50 (e.g., 2000)}

\item{exponent}{Radiator exponent (n), usually 1.3 for steel panels}
}
\value{
An object of class 'radiator'
}
\description{
Create a Radiator Object
}
