% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/room.R
\name{new_room}
\alias{new_room}
\title{Create a Room Object}
\usage{
new_room(name, volume, temp_initial = 18, u_value = 30)
}
\arguments{
\item{name}{Character name}

\item{volume}{Volume in m3}

\item{temp_initial}{Initial air temperature (°C)}

\item{u_value}{overall heat loss coefficient (W/K)}
}
\description{
Create a Room Object
}
