% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiator.R
\name{print.radiator}
\alias{print.radiator}
\title{Print method for Radiator objects}
\usage{
\method{print}{radiator}(x, ...)
}
\arguments{
\item{x}{A radiator object}

\item{...}{Additional arguments passed to print}
}
\description{
Print method for Radiator objects
}
\keyword{internal}
