% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiator.R
\name{simulate_output}
\alias{simulate_output}
\title{Calculate Heat Output}
\usage{
simulate_output(x, t_in, t_room, flow_rate)
}
\arguments{
\item{x}{A radiator object}

\item{t_in}{Inlet water temperature (°C)}

\item{t_room}{Ambient room temperature (°C)}

\item{flow_rate}{Water flow rate (kg/s or L/s)}
}
\description{
Calculate Heat Output
}
