% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_stepper.R
\name{update_room}
\alias{update_room}
\title{Advance Room Simulation}
\usage{
update_room(room, radiator, temp_outdoor, dt = 60)
}
\arguments{
\item{room}{A room object}

\item{radiator}{A radiator object}

\item{temp_outdoor}{Outdoor temperature (°C)}

\item{dt}{Time step in seconds (e.g., 60 for 1 minute)}
}
\description{
Advance Room Simulation
}
