% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/room.R
\name{update_room_multi}
\alias{update_room_multi}
\title{Update Room Temperature from Multiple Heat Sources}
\usage{
update_room_multi(room, q_total, temp_outdoor, dt = 60)
}
\arguments{
\item{room}{Room object}

\item{q_total}{Total Watts from all radiators}

\item{temp_outdoor}{Outdoor temp (°C)}

\item{dt}{Time step (seconds)}
}
\description{
Update Room Temperature from Multiple Heat Sources
}
