% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast_iron.R
\name{new_cast_iron}
\alias{new_cast_iron}
\title{Create a Cast Iron Radiator Object}
\usage{
new_cast_iron(name, height, depth, n_elements, n = 1.3)
}
\arguments{
\item{name}{Character name}

\item{height}{Height in mm}

\item{depth}{Depth in mm}

\item{n_elements}{Number of sections/columns}

\item{n}{Radiator exponent (default 1.3 for cast iron)}
}
\description{
Create a Cast Iron Radiator Object
}
